requirements.txt
boto3==1.34.106
python-dotenv==1.0.1